# contracts/openapi/core.yaml
openapi: 3.1.0
info:
  title: Ghost core API
  version: 0.1.0
  description: |
    Core Go service. Public endpoints back the web UI (and later Slack/MCP). The `internal`
    endpoints are the bounded pull-context tools the draft agent calls; every call is logged
    server-side in context_access_log under the run the bearer run token names.
servers:
  - url: http://127.0.0.1:8080
security:
  - apiToken: []
...
paths:
...
  /runs/{run_id}/decisions:
    post:
      tags: [runs]
      operationId: recordDecision
      description: |
        Same endpoint and object for web, Slack and MCP surfaces. One decision per run:
        a second submission (e.g. a Slack double-click) returns 409 with the existing decision.
        Allowed only while the run is awaiting_human.
      parameters: [{ $ref: "#/components/parameters/RunId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DecisionRequest" }
      responses:
        "201":
          description: Decision recorded
          content:
            application/json:
              schema: { $ref: "../schemas/human_decision.v1.json" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
  /runs/{run_id}/strategies:
    get:
      tags: [runs]
      operationId: getRunStrategies
      description: |
        HAR-129 section 6, 7: the strategy set of the run with the eval bundle of every candidate, exactly as
        generated. Read-only and idempotent: "View full" in Slack and the web call this; nothing here generates
        a draft. 404 `strategies_not_ready` until the set and all its bundles exist (Message 2 is only posted
        after that).
      parameters: [{ $ref: "#/components/parameters/RunId" }]
      responses:
        "200":
          description: Strategy set and the eval bundles of its candidates
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RunStrategies" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /runs/{run_id}/strategy-decision:
    get:
      tags: [runs]
      operationId: getStrategyDecision
      description: |
        HAR-136, HAR-137: the HumanStrategyDecision of the run as it stands (the choice, the saved edits, the
        send state). Read-only. Slack re-renders Message 2 from it after a modal submit, after a conflict and
        when an outcome was ambiguous. 404 `no_decision` until the human has chosen.
      parameters: [{ $ref: "#/components/parameters/RunId" }]
      responses:
        "200":
          description: The human strategy decision
          content:
            application/json:
              schema: { $ref: "../schemas/human_strategy_decision.v1.json" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    post:
      tags: [runs]
      operationId: recordStrategyDecision
      description: |
        HAR-129 section 8-10: the human chooses a candidate, and may then save edits to its recipients, subject
        and body. Choosing does not send. The first call creates the HumanStrategyDecision (201, episode status
        `chosen`) and records Ghost's preference beside the choice; later calls with the same
        `selected_candidate_id` update the pending edits (200). Core computes the `edits` diff from the
        candidate and the final artifact. 409 `selection_locked` when a different candidate is submitted after
        the choice, 409 `already_decided` after Send or Discard. Same endpoint for Slack and web.
      parameters: [{ $ref: "#/components/parameters/RunId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StrategyDecisionRequest" }
      responses:
        "200":
          description: Edits saved on the existing choice
          content:
            application/json:
              schema: { $ref: "../schemas/human_strategy_decision.v1.json" }
        "201":
          description: Choice recorded
          content:
            application/json:
              schema: { $ref: "../schemas/human_strategy_decision.v1.json" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/DecisionConflict" }
        "422": { $ref: "#/components/responses/Error" }
  /runs/{run_id}/send:
    post:
      tags: [runs]
      operationId: sendStrategy
      description: |
        HAR-129 section 10 + HAR-139: the explicit second action. `send` first re-evaluates the FINAL artifact —
        the selected candidate with the human's saved edits — through the deterministic eval engine, in the same
        transaction, at the run's replay clock. That fresh evaluation is the gate: a blocking failure on the
        final artifact refuses the send with 422 `blocking_eval` (the decision stays pending, so the human can
        edit again or discard), and an edit that repairs a previously blocking candidate sends. The send-time
        EvalResults are persisted to `eval_runs` (an unchanged artifact re-derives the same content-hashed
        results, so a retry writes nothing new). When the human changed the artifact, `send` also writes the
        HumanDelta in the same transaction: `literal_changes` computed by core, `semantic_labels` computed by
        the worker (an unreachable or misbehaving labeler degrades to an unlabeled delta — empty
        `semantic_labels`, a core-written `candidate_criterion` when unexplained — rather than failing the
        send), `unexplained` and `explained_by_eval_result_ids`
        decided by whether an eval that failed on the candidate passes on the final artifact, and the delta is
        linked back to the episode (`human_delta_id`). An unchanged send writes no delta. `send` then records
        the HumanDecision (approve when nothing was edited, edit when something was), executes the final
        artifact under the run's run mode (a dry run records the effect and sends nothing), and moves the
        episode to `decided`; the judgment inference is then generated asynchronously (Message 3). `discard`
        records a reject and evaluates nothing. Requires a prior choice (409 `no_choice`); a second call
        returns 409 `already_decided` with the existing record. Policy may also refuse a send (422), for
        example when the run is live and core has no sender (`live_send_unavailable`).
      parameters: [{ $ref: "#/components/parameters/RunId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendRequest" }
      responses:
        "200":
          description: Send decision recorded
          content:
            application/json:
              schema: { $ref: "../schemas/human_strategy_decision.v1.json" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/DecisionConflict" }
        "422": { $ref: "#/components/responses/Error" }
  /episodes/{episode_id}/judgment-inference:
    get:
      tags: [learning]
      operationId: getJudgmentInference
      description: |
        HAR-136, HAR-137: the JudgmentInference of the episode (Message 3) with the human's verdict so far.
        Read-only. 404 `inference_not_ready` until the inference has been generated after the send, and
        `not_found` for an unknown episode.
      parameters: [{ $ref: "#/components/parameters/EpisodeId" }]
      responses:
        "200":
          description: The judgment inference
          content:
            application/json:
              schema: { $ref: "../schemas/judgment_inference.v1.json" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /episodes/{episode_id}/judgment-verdict:
    post:
      tags: [learning]
      operationId: recordJudgmentVerdict
      description: |
        HAR-129 section 11, 12: the human answers Ghost's inference. `confirmed` and `corrected` set the
        verdict (corrected needs `corrected_statement`); a request with only a `note` adds the note and leaves
        the verdict as it is. The first verdict closes the episode (`judged`); later calls may add notes or
        change confirmed to corrected and are logged. Every accepted verdict and every note is also appended
        to an append-only history (`judgment_verdicts`, `judgment_notes`) — the inference object keeps the
        latest answer, the history keeps them all (HAR-139). 404 `inference_not_ready` before Message 3's
        inference exists. A correction never promotes company knowledge: it only feeds a candidate criterion.
        `no_learning` (Slack "Don't learn this") is the human's explicit opt-out: nothing is learned from the
        episode. No candidate criterion or knowledge is seeded from it (the delta seed, the verdict seed, the
        backtest support and the generator feedback of the episode all skip it) and it is terminal for learning:
        `confirmed` or `corrected` after `no_learning` is 422 `verdict_locked`, and so is `no_learning` after
        `corrected` (the correction already seeded a candidate). A repeated `no_learning` is accepted; a
        `confirmed` may still be turned into `no_learning`; notes are always accepted.
      parameters: [{ $ref: "#/components/parameters/EpisodeId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JudgmentVerdictRequest" }
      responses:
        "200":
          description: Updated inference
          content:
            application/json:
              schema: { $ref: "../schemas/judgment_inference.v1.json" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
  /accounts/{account_id}/business-intelligence/latest:
    get:
      tags: [accounts]
      operationId: getLatestBusinessIntelligence
      description: |
        HAR-129 section 5: the newest BusinessIntelligenceUpdate of the account, the object Slack Message 1
        renders and `View account map` points at. 404 `no_update` when no material change has been reported.
      parameters: [{ $ref: "#/components/parameters/AccountId" }]
      responses:
        "200":
          description: Latest business-intelligence update
          content:
            application/json:
              schema: { $ref: "../schemas/business_intelligence_update.v1.json" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    apiToken:
      type: http
      scheme: bearer
      description: Operator/UI token (GHOST_API_TOKEN). Slack and MCP adapters use their own service tokens.
    runToken:
      type: http
      scheme: bearer
      description: Run-scoped token handed to the worker in /v1/draft; binds context pulls to one run and its account.
    askToken:
      type: http
      scheme: bearer
      description: Question-scoped token handed to the worker in /v1/ask; short-lived, read tools only.
  parameters:
...
    AccountId:
      name: account_id
      in: path
      required: true
      schema: { $ref: "../schemas/common.v1.json#/$defs/uuid" }
    RunId:
      name: run_id
      in: path
      required: true
      schema: { $ref: "../schemas/common.v1.json#/$defs/uuid" }
    EpisodeId:
      name: episode_id
      in: path
      required: true
      schema: { $ref: "../schemas/common.v1.json#/$defs/uuid" }
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    DecisionConflict:
      description: |
        409 on strategy-decision and send. `already_decided` answers with the existing HumanStrategyDecision
        (nothing changed, the caller re-renders it); `selection_locked` and `no_choice` answer with the error
        envelope (code in `error.code`).
      content:
        application/json:
          schema:
            oneOf:
              - { $ref: "#/components/schemas/Error" }
              - { $ref: "../schemas/human_strategy_decision.v1.json" }
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object }
...
    DecisionRequest:
      type: object
      additionalProperties: false
      required: [decision, surface, actor_label]
      properties:
        decision: { type: string, enum: [approve, edit, reject, ignore] }
        surface: { type: string, enum: [web, slack, mcp, api] }
        actor_person_id: { oneOf: [{ $ref: "../schemas/common.v1.json#/$defs/uuid" }, { type: "null" }] }
        actor_label: { type: string, minLength: 1, maxLength: 200 }
        edited_artifact: { oneOf: [{ $ref: "../schemas/agent_run_output.v1.json#/properties/finished_artifact" }, { type: "null" }] }
        reason: { type: [string, "null"], maxLength: 2000 }
      if: { properties: { decision: { const: edit } } }
      then: { required: [edited_artifact], properties: { edited_artifact: { type: object } } }
...
    RunStrategies:
      type: object
      additionalProperties: false
      required: [strategy_set, eval_bundles]
      properties:
        strategy_set: { $ref: "../schemas/strategy_set.v1.json" }
        eval_bundles:
          description: One bundle per candidate (candidate.eval_bundle_ref points at it).
          type: array
          minItems: 3
          maxItems: 3
          items: { $ref: "../schemas/eval_bundle.v1.json" }
    StrategyDecisionRequest:
      type: object
      additionalProperties: false
      required: [selected_candidate_id, surface, actor_label]
      properties:
        selected_candidate_id: { $ref: "../schemas/common.v1.json#/$defs/uuid" }
        surface: { $ref: "../schemas/human_decision.v1.json#/properties/surface" }
        actor_person_id: { oneOf: [{ $ref: "../schemas/common.v1.json#/$defs/uuid" }, { type: "null" }] }
        actor_label: { type: string, minLength: 1, maxLength: 200 }
        final_to:
          description: Edited recipients; omit to keep the current ones (the saved edit, else the candidate's).
          type: array
          minItems: 1
          items: { $ref: "../schemas/strategy_candidate.v1.json#/properties/to/items" }
        final_cc:
          type: array
          items: { $ref: "../schemas/strategy_candidate.v1.json#/properties/cc/items" }
        final_artifact:
          description: Edited subject and body (and attachments); omit to keep the current ones (the saved edit, else the candidate's).
          $ref: "../schemas/agent_run_output.v1.json#/properties/finished_artifact"
    SendRequest:
      type: object
      additionalProperties: false
      required: [decision, surface, actor_label]
      properties:
        decision: { type: string, enum: [send, discard] }
        surface: { $ref: "../schemas/human_decision.v1.json#/properties/surface" }
        actor_person_id: { oneOf: [{ $ref: "../schemas/common.v1.json#/$defs/uuid" }, { type: "null" }] }
        actor_label: { type: string, minLength: 1, maxLength: 200 }
    JudgmentVerdictRequest:
      type: object
      additionalProperties: false
      required: [surface, actor_label]
      anyOf:
        - { required: [verdict] }
        - { required: [note] }
      properties:
        verdict: { type: string, enum: [confirmed, corrected, no_learning] }
        corrected_statement: { type: string, minLength: 1, maxLength: 1500 }
        note: { type: string, minLength: 1, maxLength: 2000 }
        surface: { $ref: "../schemas/human_decision.v1.json#/properties/surface" }
        actor_person_id: { oneOf: [{ $ref: "../schemas/common.v1.json#/$defs/uuid" }, { type: "null" }] }
        actor_label: { type: string, minLength: 1, maxLength: 200 }
      if: { required: [verdict], properties: { verdict: { const: corrected } } }
      then: { required: [corrected_statement] }
...